+++ src/main/java/egovframework/com/uss/umt/service/impl/EgovUserManageServiceImpl.java
@@ -68,16 +68,18 @@ public class EgovUserManageServiceImpl extends AbstractServiceImpl implements Eg
 	public void deleteUser(String checkedIdForDel) {
 		String [] delId = checkedIdForDel.split(",");
 		for (int i=0; i<delId.length ; i++){
-			String [] id = delId[i].split(":");
-			if (id[0].equals("USR03")){
-		        //업무사용자(직원)삭제
-				userManageDAO.deleteUser(id[1]);
-			}else if(id[0].equals("USR01")){
-				//일반회원삭제
-				mberManageDAO.deleteMber(id[1]);
-			}else if(id[0].equals("USR02")){
-				//기업회원삭제
-				entrprsManageDAO.deleteEntrprsmber(id[1]);
+			if (delId[i] != null) {
+				String [] id = delId[i].split(":");
+				if (id[0].equals("USR03")){
+			        //업무사용자(직원)삭제
+					userManageDAO.deleteUser(id[1]);
+				}else if(id[0].equals("USR01")){
+					//일반회원삭제
+					mberManageDAO.deleteMber(id[1]);
+				}else if(id[0].equals("USR02")){
+					//기업회원삭제
+					entrprsManageDAO.deleteEntrprsmber(id[1]);
+				}
 			}
 		}
 	}
